+++ dwara-db/src/data/sql/dwara_env__preprod__specific_config_values_update.sql
@@ -1,18 +1,19 @@
 SET foreign_key_checks = 0; 
 
 -- preprod env alterations
+drop table dwara_sequences;
+drop table artifact_sequence;
+drop table file_sequence;
+
 truncate artifact1;
 truncate artifact1_volume;
 truncate artifact2;
 truncate artifact2_volume;
-truncate artifact_sequence;
-truncate dwara_sequences;
 truncate file1;
 truncate file1_volume;
 truncate file2;
 truncate file2_volume;
 truncate file1_volume;
-truncate file_sequence;
 truncate job;
 truncate jobrun;
 truncate request;
@@ -27,6 +28,10 @@ update artifactclass set path_prefix = replace(path_prefix, '/dwara/', '/dwara-p
 update volume set id = replace(id, 'R', 'S');
 update volume set id = replace(id, 'G', 'H');
 update volume set id = replace(id, 'X', 'W');
+
+update volume set group_ref_id = replace(group_ref_id, 'R', 'S');
+update volume set group_ref_id = replace(group_ref_id, 'G', 'H');
+update volume set group_ref_id = replace(group_ref_id, 'X', 'W');
 
 update artifactclass_volume set volume_id = replace(volume_id, 'R', 'S');
 update artifactclass_volume set volume_id = replace(volume_id, 'G', 'H');