+++ dwara-db/src/data/sql/dwara_update_2_0_3_hotfix5.sql
@@ -1,0 +1,5 @@
+SET foreign_key_checks = 0;
+
+UPDATE `from_prd`.`artifactclass` SET `config`='{\"pathname_regex\": \"^([^/]+/?){1,2}$|^[^/]+/Output[s]?/.+.mov$\"}' WHERE `id`='video-pub';
+
+SET foreign_key_checks = 1;