+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/db/dao/transactional/TFileDao.java
@@ -1,0 +1,11 @@
+package org.ishafoundation.dwaraapi.db.dao.transactional;
+
+import java.util.List;
+
+import org.ishafoundation.dwaraapi.db.model.transactional.TFile;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TFileDao extends CrudRepository<TFile, Integer> {
+
+	List<TFile> findAllByArtifactId(int id);
+}