+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/db/model/transactional/domain/File.java
@@ -8,8 +8,10 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
 
+import org.ishafoundation.dwaraapi.db.model.transactional.FileColumns;
+
 @MappedSuperclass
-public class File {
+public class File extends FileColumns{
 
 	public static final String TABLE_NAME_PREFIX = "file";
 	
@@ -17,22 +19,22 @@ public class File {
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator = "file_sequence")
 	@Column(name="id")
 	private int id;
-	
-	// TODO - Anything > 3072 - throws Specified key was too long; max key length is 3072 bytes and doesnt set the unique constraint
-	@Column(length=3072, name="pathname", unique=true)
-	private String pathname;
-
-	@Column(name="checksum")
-	private byte[] checksum;
-
-	@Column(name="size")
-	private long size;
-
-	@Column(name="deleted")
-	private boolean deleted;
-	
-	@Column(name="directory")
-	private boolean directory;
+//	
+//	// TODO - Anything > 3072 - throws Specified key was too long; max key length is 3072 bytes and doesnt set the unique constraint
+//	@Column(length=3072, name="pathname", unique=true)
+//	private String pathname;
+//
+//	@Column(name="checksum")
+//	private byte[] checksum;
+//
+//	@Column(name="size")
+//	private long size;
+//
+//	@Column(name="deleted")
+//	private boolean deleted;
+//	
+//	@Column(name="directory")
+//	private boolean directory;
 
 	public int getId() {
 		return id;
@@ -42,56 +44,56 @@ public class File {
 		this.id = id;
 	}
 
-	public String getPathname() {
-		return pathname;
-	}
-
-	public void setPathname(String pathname) {
-		this.pathname = pathname;
-	}
-
-	public byte[] getChecksum() {
-		return checksum;
-	}
-
-	public void setChecksum(byte[] checksum) {
-		this.checksum = checksum;
-	}
-
-	public long getSize() {
-		return size;
-	}
-
-	public void setSize(long size) {
-		this.size = size;
-	}
-
-	public boolean isDeleted() {
-		return deleted;
-	}
-
-	public void setDeleted(boolean deleted) {
-		this.deleted = deleted;
-	}
-
-	public boolean isDirectory() {
-		return directory;
-	}
-
-	public void setDirectory(boolean directory) {
-		this.directory = directory;
-	}
+//	public String getPathname() {
+//		return pathname;
+//	}
+//
+//	public void setPathname(String pathname) {
+//		this.pathname = pathname;
+//	}
+//
+//	public byte[] getChecksum() {
+//		return checksum;
+//	}
+//
+//	public void setChecksum(byte[] checksum) {
+//		this.checksum = checksum;
+//	}
+//
+//	public long getSize() {
+//		return size;
+//	}
+//
+//	public void setSize(long size) {
+//		this.size = size;
+//	}
+//
+//	public boolean isDeleted() {
+//		return deleted;
+//	}
+//
+//	public void setDeleted(boolean deleted) {
+//		this.deleted = deleted;
+//	}
+//
+//	public boolean isDirectory() {
+//		return directory;
+//	}
+//
+//	public void setDirectory(boolean directory) {
+//		this.directory = directory;
+//	}
 
 	@Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         File file = (File) o;
-        return Objects.equals(pathname, file.pathname);
+        return Objects.equals(getPathname(), file.getPathname());
     }
  
     @Override
     public int hashCode() {
-        return Objects.hash(pathname);
+        return Objects.hash(getPathname());
     }
 }